+++ client/src/App.css
@@ -179,3 +179,11 @@ li {
   background-color: #b6382d;
   color: white;
 }
+
+.alert-wrapper {
+  position: fixed;
+  top: 4rem;
+  right: 2rem;
+  display: inline-block;
+  z-index: 3;
+}
