+++ client/src/components/Alert.jsx
@@ -1,0 +1,22 @@
+import React from "react";
+import { useSelector } from "react-redux";
+
+const Alert = () => {
+  // App State
+  const alert = useSelector((state) => state.alert);
+  return (
+    <div className="alert-wrapper">
+      {alert.map((alert) => (
+        <div
+          key={alert.id}
+          className={`alert alert-${alert.alertType}`}
+          style={{ fontFamily: "sans-serif" }}
+        >
+          {alert.message}
+        </div>
+      ))}
+    </div>
+  );
+};
+
+export default Alert;
