+++ client/src/constants/alertConstants.js
@@ -1,0 +1,2 @@
+export const SET_ALERT = "SET_ALERT";
+export const REMOVE_ALERT = "REMOVE_ALERT";
