+++ server/routes/supportRoutes.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { sendSupportMail } from "../controllers/supportController.js";
+import { sendSupportMail } from "../controllers/Support/supportController.js";
 
 const router = express.Router();
