+++ server/routes/todoRoutes.js
@@ -1,10 +1,8 @@
 import express from "express";
-import {
-  addListItem,
-  createList,
-  getTodoLists,
-  removeList,
-} from "../controllers/todoController.js";
+import { getTodoLists } from "../controllers/Todo/getTodoLists.js";
+import { createList } from "../controllers/Todo/createList.js";
+import { addListItem } from "../controllers/Todo/addListItem.js";
+import { removeList } from "../controllers/Todo/removeList.js";
 import { protect, authorize } from "../middleware/auth.js";
 
 const router = express.Router();
